changed the project structure
