+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules/
+venv/
 dist/
 .cache/
 __pycache__/
