+++ src/todo_list/project.py
@@ -1,0 +1,10 @@
+from json_record import JSONRecord
+
+class Project(JSONRecord):
+    def __init__(self):
+        self.project_breakdown = {
+            "project": "",
+            "sub_project": "",
+        }
+        super().__init__(self.project_breakdown)
+    # todo
